also extract out the utility class
